+++ instance/instance_test.go
@@ -10,7 +10,7 @@ import (
 var (
 	fixedTime    = time.Unix(1000000000, 0)
 	exampleImage = "owner/repo:tag"
-	testRegistry = NewMockRegistry([]flux.ImageDescription{
+	testRegistry = registry.NewMockRegistry([]flux.ImageDescription{
 		{
 			ID:        flux.ParseImageID(exampleImage),
 			CreatedAt: &fixedTime,
@@ -31,25 +31,9 @@ func TestSomething(t *testing.T) {
 func testImageExists(t *testing.T, i Instance, image string, expected bool) {
 	b, err := i.imageExists(flux.ParseImageID(image))
 	if err != nil {
-		t.Fatal(err.Error())
+		t.Fatalf("%v: error when requesting image %q", err.Error(), image)
 	}
 	if b != expected {
 		t.Fatalf("Expected exist = %q but got %q", expected, b)
 	}
-}
-
-type mockRegistry struct {
-	descriptions []flux.ImageDescription
-	err          error
-}
-
-func NewMockRegistry(descriptions []flux.ImageDescription, err error) registry.Client {
-	return &mockRegistry{
-		descriptions: descriptions,
-		err:          err,
-	}
-}
-
-func (r *mockRegistry) GetRepository(repository string) ([]flux.ImageDescription, error) {
-	return r.descriptions, r.err
 }
